lambdas/html_body/requirements.txt
psycopg2-binary 
python-dotenv
boto3